# backend/api/openapi.yaml
openapi: 3.0.3
info:
  title: Climate Invest Optimizer API
  description: |
    API REST para la optimización de presupuestos de inmuebles orientada a la prevención 
    y mitigación de riesgos climáticos.
    
    ## Características principales
    
    - **Gestión de inmuebles (Shops)**: CRUD completo de tiendas con información de localización y riesgos
    - **Evaluación de riesgos climáticos**: Análisis de exposición, sensibilidad y probabilidad por cluster
    - **Medidas preventivas**: Catálogo de medidas con costos estimados y tipos
    - **Optimización de presupuesto**: Algoritmos (greedy, knapsack, weighted) para maximizar reducción de riesgo
    - **Dashboard**: Estadísticas y métricas generales
    
    ## Autenticación
    
    La API utiliza autenticación JWT. Incluye el token en el header `Authorization`:
    ```
    Authorization: Bearer <token>
    ```
    
    ## Códigos de error
    
    | Código | Descripción |
    |--------|-------------|
    | 400 | Bad Request - Error de validación |
    | 401 | Unauthorized - Token inválido o expirado |
    | 403 | Forbidden - Sin permisos suficientes |
    | 404 | Not Found - Recurso no encontrado |
    | 409 | Conflict - Recurso duplicado |
    | 422 | Unprocessable Entity - Error de negocio |
    | 500 | Internal Server Error - Error del servidor |
    
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
...
paths:
...
  /shops:
    get:
      tags:
        - shops
      summary: Lista todas las tiendas
      description: Obtiene una lista paginada de tiendas con filtros opcionales
      operationId: listShops
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
        - name: cluster_id
          in: query
          description: Filtrar por ID de cluster
          schema:
            type: integer
        - name: min_risk
          in: query
          description: Riesgo mínimo
          schema:
            type: number
            format: float
        - name: max_risk
          in: query
          description: Riesgo máximo
          schema:
            type: number
            format: float
        - name: q
          in: query
          description: Búsqueda por localización
          schema:
            type: string
      responses:
        '200':
          description: Lista de tiendas
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/PaginatedShops'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - shops
      summary: Crear tienda
      description: Crea una nueva tienda
      operationId: createShop
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateShopRequest'
      responses:
        '201':
          description: Tienda creada
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Shop'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /shops/{id}:
    get:
      tags:
        - shops
      summary: Obtener tienda por ID
      description: Retorna los detalles completos de una tienda
      operationId: getShop
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ShopIdParam'
      responses:
        '200':
          description: Detalles de la tienda
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/ShopWithDetails'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      tags:
        - shops
      summary: Actualizar tienda
      description: Actualiza parcialmente una tienda
      operationId: updateShop
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ShopIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateShopRequest'
      responses:
        '200':
          description: Tienda actualizada
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Shop'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags:
        - shops
      summary: Eliminar tienda
      description: Elimina una tienda y sus medidas asociadas
      operationId: deleteShop
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ShopIdParam'
      responses:
        '204':
          description: Tienda eliminada
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /shops/{id}/measures:
    post:
      tags:
        - shops
      summary: Aplicar medidas a tienda
      description: Aplica una o más medidas preventivas a una tienda
      operationId: applyMeasures
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ShopIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApplyMeasuresRequest'
      responses:
        '200':
          description: Medidas aplicadas
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /shops/{id}/risk-assessment:
    get:
      tags:
        - shops
      summary: Evaluación de riesgos
      description: Obtiene la evaluación de riesgos climáticos de una tienda
      operationId: getRiskAssessment
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ShopIdParam'
      responses:
        '200':
          description: Evaluación de riesgos
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/RiskAssessment'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token JWT obtenido del endpoint /auth/login
  parameters:
    ShopIdParam:
      name: id
      in: path
      required: true
      description: ID de la tienda
      schema:
        type: integer
        format: int64
    PageParam:
      name: page
      in: query
      description: Número de página
      schema:
        type: integer
        default: 1
        minimum: 1
    PageSizeParam:
      name: page_size
      in: query
      description: Elementos por página
      schema:
        type: integer
        default: 20
        minimum: 1
        maximum: 100
  responses:
    BadRequest:
      description: Error de validación
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: VALIDATION_ERROR
              message: "El campo 'location' es requerido"
    Unauthorized:
      description: No autorizado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: UNAUTHORIZED
              message: Token de autenticación requerido
...
    NotFound:
      description: Recurso no encontrado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: NOT_FOUND
              message: Tienda no encontrada
...
  schemas:
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Operación exitosa
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              example: ERROR_CODE
            message:
              type: string
              example: Descripción del error
...
    Shop:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        location:
          type: string
          example: Barcelona, Spain
        utm_north:
          type: number
          format: float
          example: 41.3851
        utm_east:
          type: number
          format: float
          example: 2.1734
        total_risk:
          type: number
          format: float
          example: 0.65
        taxonomy_coverage:
          type: number
          format: float
          example: 45.5
        surface:
          type: number
          format: float
          example: 2500
        carbon_footprint:
          type: number
          format: float
          example: 150.5
        cluster_id:
          type: integer
          format: int64
          example: 3
    ShopWithDetails:
      allOf:
        - $ref: '#/components/schemas/Shop'
        - type: object
          properties:
            cluster_name:
              type: string
              example: Mediterranean Coast
            risks:
              type: array
              items:
                $ref: '#/components/schemas/RiskDetail'
            applied_measures:
              type: array
              items:
                $ref: '#/components/schemas/Measure'
    CreateShopRequest:
      type: object
      required:
        - location
        - utm_north
        - utm_east
        - surface
        - cluster_id
      properties:
        location:
          type: string
          minLength: 3
          maxLength: 255
          example: Madrid, Spain
        utm_north:
          type: number
          format: float
          example: 40.4168
        utm_east:
          type: number
          format: float
          example: -3.7038
        surface:
          type: number
          format: float
          minimum: 0
          example: 3000
        carbon_footprint:
          type: number
          format: float
          minimum: 0
          example: 120
        cluster_id:
          type: integer
          format: int64
          example: 2
    UpdateShopRequest:
      type: object
      properties:
        location:
          type: string
          minLength: 3
          maxLength: 255
        utm_north:
          type: number
          format: float
        utm_east:
          type: number
          format: float
        surface:
          type: number
          format: float
          minimum: 0
        carbon_footprint:
          type: number
          format: float
          minimum: 0
        cluster_id:
          type: integer
          format: int64
    PaginatedShops:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Shop'
        pagination:
          $ref: '#/components/schemas/PaginationMeta'
    PaginationMeta:
      type: object
      properties:
        page:
          type: integer
          example: 1
        page_size:
          type: integer
          example: 20
        total_items:
          type: integer
          format: int64
          example: 150
        total_pages:
          type: integer
          example: 8
        has_next:
          type: boolean
          example: true
        has_prev:
          type: boolean
          example: false
...
    Risk:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        name:
          type: string
          example: Inundación costera/fluvial/pluvial
        svg:
          type: string
          nullable: true
    RiskDetail:
      allOf:
        - $ref: '#/components/schemas/Risk'
        - type: object
          properties:
            exposure:
              type: string
              enum: [very_low, low, medium, high, very_high]
              example: medium
            sensitivity:
              type: string
              enum: [very_low, low, medium, high, very_high]
              example: high
            consequence:
              type: string
              enum: [very_low, low, medium, high, very_high]
              example: high
            probability:
              type: string
              enum: [very_low, low, medium, high, very_high]
              example: medium
            risk_score:
              type: number
              format: float
              example: 0.42
    Measure:
      type: object
      properties:
        name:
          type: string
          example: Sistema de drenaje perimetral
        estimated_cost:
          type: number
          format: float
          example: 4500
        type:
          type: string
          enum: [natural, material, immaterial]
          example: material
    ApplyMeasuresRequest:
      type: object
      required:
        - measure_names
      properties:
        measure_names:
          type: array
          minItems: 1
          items:
            type: string
          example: ["Sistema de drenaje perimetral", "Aislamiento térmico continuo"]
    RiskAssessment:
      type: object
      properties:
        shop_id:
          type: integer
          format: int64
        overall_risk_score:
          type: number
          format: float
          example: 0.65
        risk_level:
          type: string
          enum: [very_low, low, medium, high, very_high]
          example: medium
        risks:
          type: array
          items:
            $ref: '#/components/schemas/RiskDetail'
        last_updated:
          type: string
          format: date-time
...